Move test from db_test.go to value_test.go
