+++ db_test.go
@@ -36,7 +36,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/require"
-	"golang.org/x/net/trace"
 
 	"github.com/dgraph-io/badger/options"
 	"github.com/dgraph-io/badger/pb"
@@ -1894,55 +1893,4 @@ func ExampleDB_Subscribe() {
 	wg.Wait()
 	// Output:
 	// a-key is now set to a-value
-}
-
-// Regression test for https://github.com/dgraph-io/badger/issues/926
-func TestDiscardStatsMove(t *testing.T) {
-	dir, err := ioutil.TempDir("", "badger-test")
-	require.NoError(t, err)
-	ops := getTestOptions(dir)
-	ops.ValueLogMaxEntries = 1
-	db, err := Open(ops)
-	require.NoError(t, err)
-
-	stat := make(map[uint32]int64, ops.ValueThreshold+10)
-	for i := uint32(0); i < uint32(ops.ValueThreshold+10); i++ {
-		stat[i] = 0
-	}
-
-	// Set discard stat.
-	db.vlog.lfDiscardStats = &lfDiscardStats{
-		m: stat,
-	}
-	entries := []*Entry{{
-		Key: y.KeyWithTs(lfDiscardStatsKey, 1),
-		// The discard stat value is more than value threshold.
-		Value: db.vlog.encodedDiscardStats(),
-	}}
-	// Push discard stats entry to write channel
-	req, err := db.sendToWriteCh(entries)
-	require.NoError(t, err)
-	req.Wait()
-
-	// Push more entries so that we get more than 1 value log files.
-	require.NoError(t, db.Update(func(txn *Txn) error {
-		e := NewEntry([]byte("f"), []byte("1"))
-		return txn.SetEntry(e)
-	}))
-	require.NoError(t, db.Update(func(txn *Txn) error {
-		e := NewEntry([]byte("ff"), []byte("1"))
-		return txn.SetEntry(e)
-
-	}))
-
-	tr := trace.New("Badger.ValueLog", "GC")
-	// Use first value log file for GC. This value log file contains the discard stats.
-	lf := db.vlog.filesMap[0]
-	require.NoError(t, db.vlog.rewrite(lf, tr))
-	require.NoError(t, db.Close())
-
-	db, err = Open(ops)
-	require.NoError(t, err)
-	require.Equal(t, stat, db.vlog.lfDiscardStats.m)
-	require.NoError(t, db.Close())
 }
